OK, new plan is to rip all of the vivado project files and generated stuff of the repository, and to work from the Xillybus distribution, which generally seems to follow best practice for revision control with Vivado.  Instead of checking in the Vivado project, you check in a Vivado TCL script which recreates the project.  Since xillybus is the only non-Xilinx IP we are using, this seems like it should be pretty straightforward.
In order to be resilient with respect to whatever Xillybus may do in the future. I am checking in their binary distribution.  The main vulnerability there is that this version of the IP might stop working with future versions of Vivado.  You unzip this, and then a modified version of their setup script reaches into that distribution and sets things up as they have done, but also adds our new sources.

I haven't yet actually implemented this scheme yet, this commit just deletes all of the Vivado project stuff and adds the Xillybus distribution.